file_cache: Load templates from any web-root/filesystem
I think the ideally we'd check modification time first, check that
the first valid file matches the _source filesystem_ of the cache entry
(which we'd have to store), otherwise load the higher priority file,
and have a config option for whether or not to fall-through to a lower
priority filesystem if the highest-priority non-missing modification
time (or the file backing it) errors on-read.
Tests would also be great.

But since this meets my needs for the time being, and (IIUC) matches the
functionality of existing filesystem -> database -> cache layers, I'll
have to put a pin in it for now.

* src/file_cache.rs: Load templates from any web-root/filesystem